+++ R/internals.R
@@ -39,3 +39,53 @@ get_symbols = function(call)
 
   return(unique(csymbols))
 }
+
+# nocov start
+install_cmd_tools = function()
+{
+  get_os <- function(){
+    sysinf <- Sys.info()
+    if (!is.null(sysinf)){
+      os <- sysinf['sysname']
+      if (os == 'Darwin')
+        os <- "osx"
+    } else { ## mystery machine
+      os <- .Platform$OS.type
+      if (grepl("^darwin", R.version$os))
+        os <- "osx"
+      if (grepl("linux-gnu", R.version$os))
+        os <- "linux"
+    }
+    tolower(os)
+  }
+
+  f <- system.file("lasr.R", package = "lasR")
+  d <- system.file("", package = "lasR")
+
+  if (o == "linux") {
+    cmd = paste("ln -s", f, "~/.local/bin/lasr")
+    system(cmd)
+    return(invisible())
+  }
+
+  if (os == "windows") {
+    bat = substr(f, 1, nchar(f)-1)
+    batch_file_path = paste0(bat, "bat")
+    batch_content <- c("@echo off", paste("Rscript", f, "%*"))
+    writeLines(batch_content, con = batch_file_path)
+
+    cat("To run the 'lasr' command from anywhere in the Command Prompt, you need to add its location to your system's PATH:
+
+  1. Right-click on This PC or My Computer and select Properties.
+  2. Click on Advanced system settings and then Environment Variables.
+  3. In the System variables section, find the Path variable, select it, and click Edit.
+  4. Add this path:", d)
+
+    return(invisible())
+  }
+
+  if (os == "osx") stop("MacOS not supported")
+
+  stop("We failed to detect your OS")
+}
+# nocov end